# Faster update max quantities using NOWAIT