--- a/.gitignore
+++ b/.gitignore
@@ -7,6 +7,7 @@ propagation/SWALS/tests/*/*.log
 propagation/SWALS/tests/*/*unit_tests
 propagation/SWALS/tests/unit_tests/*.csv
 propagation/SWALS/tests/unit_tests/*.nc
+propagation/SWALS/tests/validation_tests/*.RDS
 
 # Example outputs
 # NB: a/**/b matches a/b, a/x/b, a/x/y/b, ...