Demonstration of the small improvement on example benchmark problem nthmp/BP09 with openmp. Add the DEVOLVE_TIMER flag to make BP09 to track the update_max_quantities time.

On an NCI node with two Cascade Lakes processors (48 cores) using 48 threads took: the run takes 1m25s with the update_max_quantities taking a total of 1.02s over all domains. Without the NOWAIT (the original), it takes 1m30s but update_max_quantities takes a total of 7.17s.