@gareth-d-ga 2024-10-16
Thanks Matt. That timing difference is interesting, noting the improved performance is better than on my machine with 16 cores (where the improvement was marginal).

I suspect the difference is that here you are using 48 cores over 2 CPUs -- which is probably inefficient (too many cores, slow to synchronize between CPUs). So the overheads of synchronisation are larger, and NOWAIT improves this. 

Can you check whether all the unit/parallel/validation tests PASS on NCI, and add a comment here.