@matt-mac-ga 2024-10-16
The most recent timing available at on NCI at ptha_mm/propagation/SWALS/examples/nthmp/BP09/OUTPUTS/openmp_results/RUN_ID*/Evolve_timer_details.txt.

All unit tests, parallel tests and validation tests on NCI using ifx (intel-compiler-llvm/2024.0.2).