@gareth-d-ga 2024-10-16
> The most recent timing available at on NCI at ptha_mm/propagation/SWALS/examples/nthmp/BP09/OUTPUTS/openmp_results/RUN_ID*/Evolve_timer_details.txt.

But that's only one model -- to see the speedup I need the other model as well.

> All unit tests, parallel tests and validation tests on NCI using ifx (intel-compiler-llvm/2024.0.2).

Great, thanks