@matt-mac-ga 2024-10-16
The results of the evolve timer are below for one and two cascade lakes CPUs on NCI. Only the line with update_max_quantities are shown.

# Using 48 threads on 2 Cascade lakes CPUS (24 cores each. 48 cores total)

## With NOWAIT
```
BP09/OUTPUTS_NOWAIT/openmp_results/RUN_ID00000000010000000001_00001_20241016_175344.716/Evolve_timer_details.txt:                                  update_max_quantities:           0.11152411 :  14.9288561807499516 %
BP09/OUTPUTS_NOWAIT/openmp_results/RUN_ID00000000010000000002_00001_20241016_175344.799/Evolve_timer_details.txt:                                  update_max_quantities:           0.14983273 :   1.5170936529236070 %
BP09/OUTPUTS_NOWAIT/openmp_results/RUN_ID00000000010000000003_00001_20241016_175344.809/Evolve_timer_details.txt:                                  update_max_quantities:           0.23244858 :   1.3181878200235022 %
BP09/OUTPUTS_NOWAIT/openmp_results/RUN_ID00000000010000000004_00001_20241016_175344.820/Evolve_timer_details.txt:                                  update_max_quantities:           0.49302626 :   1.3871461536397542 %
BP09/OUTPUTS_NOWAIT/openmp_results/RUN_ID00000000010000000005_00001_20241016_175344.831/Evolve_timer_details.txt:                                  update_max_quantities:           0.43365812 :   1.5213948519220979 %
BP09/OUTPUTS_NOWAIT/openmp_results/RUN_ID00000000010000000006_00001_20241016_175344.842/Evolve_timer_details.txt:                                  update_max_quantities:           0.17146587 :   1.5165463988591676 %
```
**Totalling  1.53 seconds**


## Original
```
BP09/OUTPUTS/RUN_20241017_101700664/RUN_ID00000000010000000001_00001_20241017_101700.690/Evolve_timer_details.txt:                                  update_max_quantities:           0.52720428 :  50.3622085119398832 %
BP09/OUTPUTS/RUN_20241017_101700664/RUN_ID00000000010000000002_00001_20241017_101700.809/Evolve_timer_details.txt:                                  update_max_quantities:           0.41463065 :   8.6800162531287537 %
BP09/OUTPUTS/RUN_20241017_101700664/RUN_ID00000000010000000003_00001_20241017_101700.819/Evolve_timer_details.txt:                                  update_max_quantities:           0.80468893 :   9.7762751352152382 %
BP09/OUTPUTS/RUN_20241017_101700664/RUN_ID00000000010000000004_00001_20241017_101700.829/Evolve_timer_details.txt:                                  update_max_quantities:           2.27008915 :  13.2828868509569986 %
BP09/OUTPUTS/RUN_20241017_101700664/RUN_ID00000000010000000005_00001_20241017_101700.839/Evolve_timer_details.txt:                                  update_max_quantities:           2.32767177 :  15.9781675507472816 %
BP09/OUTPUTS/RUN_20241017_101700664/RUN_ID00000000010000000006_00001_20241017_101700.850/Evolve_timer_details.txt:                                  update_max_quantities:           0.77818060 :  14.3665670432987511 %
```
**Totalling 7.12 seconds**

# Using 24 threads on 1 cascade lake cpu (24 cores)
## With NOWAIT
```
OUTPUTS/RUN_20241017_103605380/RUN_ID00000000010000000001_00001_20241017_103605.393/Evolve_timer_details.txt:                                  update_max_quantities:           0.10040998 :  19.9957363952788896 %
OUTPUTS/RUN_20241017_103605380/RUN_ID00000000010000000002_00001_20241017_103605.438/Evolve_timer_details.txt:                                  update_max_quantities:           0.15944338 :   2.7262800985637616 %
OUTPUTS/RUN_20241017_103605380/RUN_ID00000000010000000003_00001_20241017_103605.447/Evolve_timer_details.txt:                                  update_max_quantities:           0.19077873 :   1.9087289216546526 %
OUTPUTS/RUN_20241017_103605380/RUN_ID00000000010000000004_00001_20241017_103605.456/Evolve_timer_details.txt:                                  update_max_quantities:           0.35133982 :   1.8493088135742495 %
OUTPUTS/RUN_20241017_103605380/RUN_ID00000000010000000005_00001_20241017_103605.465/Evolve_timer_details.txt:                                  update_max_quantities:           0.31546521 :   1.9385092558107977 %
OUTPUTS/RUN_20241017_103605380/RUN_ID00000000010000000006_00001_20241017_103605.474/Evolve_timer_details.txt:                                  update_max_quantities:           0.15878534 :   2.4922536429049598 %
```
**Totalling  1.28 seconds**

## Original
```
OUTPUTS/RUN_20241017_103016872/RUN_ID00000000010000000001_00001_20241017_103016.889/Evolve_timer_details.txt:                                  update_max_quantities:           0.36639762 :  47.7200446899063309 %
OUTPUTS/RUN_20241017_103016872/RUN_ID00000000010000000002_00001_20241017_103016.946/Evolve_timer_details.txt:                                  update_max_quantities:           0.35819292 :   5.9860615693707704 %
OUTPUTS/RUN_20241017_103016872/RUN_ID00000000010000000003_00001_20241017_103016.955/Evolve_timer_details.txt:                                  update_max_quantities:           0.57209611 :   5.5316430559805614 %
OUTPUTS/RUN_20241017_103016872/RUN_ID00000000010000000004_00001_20241017_103016.964/Evolve_timer_details.txt:                                  update_max_quantities:           1.46821642 :   7.3958245111173246 %
OUTPUTS/RUN_20241017_103016872/RUN_ID00000000010000000005_00001_20241017_103016.973/Evolve_timer_details.txt:                                  update_max_quantities:           1.43839574 :   8.3675849541142444 %
OUTPUTS/RUN_20241017_103016872/RUN_ID00000000010000000006_00001_20241017_103016.982/Evolve_timer_details.txt:                                  update_max_quantities:           0.53696203 :   8.0523198317472886 %
```
**Totalling 4.74 seconds**

The speed up is more significant over two CPUs (5.6 seconds), but still 3.5 seconds on a single CPU.